Add files via upload
Worked with team of three people to build a Peg Game that adheres to software engineering principles.  

Created a PTUI to take user input, and game can be played using the command line. Created a backtracking algorithm to suggest next possible move, along with auto-solve functionality that uses backtracking to find optimum solution.  

Main role was scrum master to coordinate team activities, lead programmer to develop the backtracking algorithm, specifically the pruning and storing of successors.